ext/dec: Add sequence of activities
Sequence of activities can be declared into the decision plan
like in the example below. It produces a set of homogeneous activities
(same rules and functions) {"id_1","id_2",...} of size 'number'
which can be infinite. For all i a constraint FS(id_i, id_{i+1})
is added. Activities are dynamically added when the current activity
is done.

activities{
  sequence-activity{
    id-prefix= "id";
    number = 2;          #optional
    rules = "", "", "";  #optional
    rules-fail = "", ""; #optional
    temporal {
      start = 10;
      finish = 20;
      minstart = 9;
      maxstart = 11;
      minfinish = 19;
      maxfinish = 21;
    }
    temporal-sequence {
      precedence {
        type = FS;
        mintemilag = 1;
        maxtimelag = 3;
      }
    }
    ack = "";
    output = "";
  }
}